Extended the HEPMC3 format to the two photon cases.
This includes both the GammaGammaSingle cases and the GammaGammaleptonpair cases. Also I fixed an error in the definintion of the getter and setter of the mass of the starlightParticle Object. I ensured the mass is specified as declared as a double presicion number and not anymore short integer.
	modified:   include/gammagammaleptonpair.h
	modified:   include/gammagammasingle.h
	modified:   include/starlightparticle.h
	modified:   src/gammaavm.cpp
	modified:   src/gammagammaleptonpair.cpp
	modified:   src/gammagammasingle.cpp

+++ include/starlightparticle.h
@@ -54,8 +54,8 @@ class starlightParticle : public lorentzVector
       void setCharge(short charge) { _charge = charge; }
       short getCharge() const { return _charge; }
 
-      void setMass(short mass) { _mass = mass; }
-      short getMass() const { return _mass; }
+      void setMass(double mass) { _mass = mass; }
+      double getMass() const { return _mass; }
       
       void setFirstParent(int parent) { _firstParent = parent; }
       void setLastParent(int parent) { _lastParent = parent; }
